+++ ArthasEngine/ArthasEngine/AEPoint.h
@@ -19,31 +19,31 @@ class AEPoint
 {
 public:
 	AEPoint();
-	AEPoint(double _x, double _y);
+	AEPoint(float _x, float _y);
 	AEPoint( const AEPoint& _point );
 	~AEPoint();
 	
-	double	GetPosX() const {return m_PointX;}
-	double	GetPosY() const {return m_PointY;}
-	void	SetPosX( double _x );
-	void	SetPosY( double _y );
-	void	SetPoint( double _x , double _y );
+	float	GetPosX() const {return m_PointX;}
+	float	GetPosY() const {return m_PointY;}
+	void	SetPosX( float _x );
+	void	SetPosY( float _y );
+	void	SetPoint( float _x , float _y );
 
 	AEPoint	operator+( const AEPoint& _point );
 	void	operator=( const AEPoint& _point );
 	bool	operator==( const AEPoint& _point );
 
 	//다른 점과의 거리를 리턴
-	double	GetDistance( const AEPoint& _point );
+	float	GetDistance( const AEPoint& _point );
 	//다른 점과의 중점을 리턴
 	AEPoint	GetCenter( const AEPoint& _point );
 	//다른 점과의 각도(라디안)를 리턴
-	double	GetAngleRadian( const AEPoint& _point );
+	float	GetAngleRadian( const AEPoint& _point );
 	//다른 점과의 각도(Degree)를 리턴
-	double	GetAngleDegree( const AEPoint& _point );
+	float	GetAngleDegree( const AEPoint& _point );
 
 
 private:
-	double m_PointX;
-	double m_PointY;
+	float m_PointX;
+	float m_PointY;
 };
